[refactor] SvgImage.js
